requirements.txt
web3
PyYAML
matplotlib
Flask
requests_toolbelt